+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## UNRELEASED
+* ideas-on-map: make the introblock optional
+* ideas-on-map: make the location marker configurable
+* ideas-on-map: search adresses municipality is now configurable
+* ideas-on-map: fix first click on map bug
 * Add login required to choices-guide results
 * Refactor